feat: Оптимизирована сборка бандла
